requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.0.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
exceptiongroup==1.3.0
faster-whisper==1.1.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
openai==1.99.0
packaging==25.0
pluggy==1.6.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3